+++ CHANGELOG.rst
@@ -6,6 +6,17 @@ HEAD
 
 **To be included in next release**
 
+0.13 (2018-03-27)
+-----------------
+
+`commit: 9642c1dcd7ba39b91ef791039690f5be79312dd2`
+
+**Add SystemTap based complexity collector**
+
+  - add SystemTap based complexity collector
+  - add ``perun utils create`` command (see :ref:`cli-utils-ref` for more details) for creating new modules according to stored templates
+  - fix issue with getting config hierarchy, when outside of any perun scope
+
 0.12.1 (2018-03-08)
 -------------------
 
@@ -15,7 +26,6 @@ HEAD
 
   - update the project readme
   - add compiled documentation
-  -
 
 0.12 (2018-03-05)
 -----------------